# Optimize DateTime constructors (DateToTicks and IsLeapYear)