Just a nano-optimization to make some of the `DateTime` constructors a bit faster, e.g.:
```csharp
[Benchmark]
public DateTime CreateDate_const()
{
    return new DateTime(2021, 5, 20);
}

[Benchmark]
[Arguments(2021, 5, 20)]
public DateTime CreateDate_var(int y, int m, int d)
{
    return new DateTime(y, m, d); // and other constructors with DateToTicks/IsLeapYear checks inside
}

/*

            |           Method |      Mean |
            |----------------- |----------:|
     master | CreateDate_const | 0.3712 ns |
     PR     | CreateDate_const | 0.2356 ns |
     master |   CreateDate_var | 2.6116 ns |
     PR     |   CreateDate_var | 2.1922 ns |

*/
```
Codegen diff for `CreateDate_var`: https://www.diffchecker.com/BOTFFqql (336 bytes -> 286 bytes), for arm64: https://www.diffchecker.com/H8xUCOtO (448 bytes -> 400 bytes)

Feel free to close if you think the hacks aren't worth it 🙂